backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
xgboost==2.1.1
openai==1.40.0
python-dotenv==1.0.1
joblib==1.4.2
gunicorn==22.0.0
